refactor: Criando molecula para o messages menu
